+++ src/core/state/root.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
 import assessment from '@covid/core/assessment/state/reducers';
-import { contentSlice, ContentState } from '@covid/core/content/state/slices';
+import { contentSlice, ContentState } from '@covid/core/content/state/contentSlice';
 import { IAssessmentState } from '@covid/core/assessment/AssessmentState';
 
 export type RootState = {
